requirements.txt
# created by pip list from local virtual environment
anyio              == 3.6.2
cryptography       == 37.0.4
certifi            == 2022.9.24
charset-normalizer == 2.1.1
click              == 8.1.3
colorama           ==  0.4.6
dnspython          == 2.2.1
email-validator    == 1.3.0
fastapi            == 0.87.0
greenlet           == 2.0.1
h11                == 0.14.0
idna               == 3.4
passlib            == 1.7.4
pydantic           == 1.10.2
PyJWT              == 2.6.0
requests           == 2.28.1
sniffio            == 1.3.0
SQLAlchemy         == 1.4.44
starlette          == 0.21.0
typing_extensions  == 4.4.0
urllib3            == 1.26.12
uvicorn            == 0.19.0
wincertstore       == 0.2
psycopg2-binary    == 2.9.5
python-dotenv      == 0.21.0
google-api-python-client == 2.86.0
google-cloud-logging == 3.5.0
pytest == 7.3.1
httpx == 0.24.0